Files reorganization and rename project to AFT
