+++ conf.d/controller/etc/aft-mapis.json
@@ -4,7 +4,7 @@
     "metadata": {
         "uid": "Test",
         "version": "1.0",
-        "api": "test",
+        "api": "afTest",
         "info": "Configuration to test an emulated api",
         "require": [
             "low-can"
