--- conf.d/project/htdocs/README.md
@@ -1,7 +1,0 @@
-------------------------------------------------------------------------
-   Basic HTML & WS test
-------------------------------------------------------------------------
-
-  # Load bindings directly from development tree for debug
-  AFB_BINDER_NAME='sample' afb-daemon --verbose --verbose --token="" --ldpaths=build --port=1234 --roothttp=htdocs
-
